Dynamic life update

+++ PixRunner/src/StaticThings.java
@@ -27,7 +27,7 @@ public class StaticThings {
         return imageView;
     }
 
-    public void update(double posX) {
+    public void update(long posX) {
         //Giving a new position for the background
         sizeX = posX;
         this.imageView.setX(sizeX);
